libmx

+++ inc/libmx.h
@@ -1,6 +1,7 @@
 #ifndef LIBMX_H
 #define LIBMX_H
 
+#include <stdio.h>
 #include <unistd.h>
 #include <wchar.h>
 #include <stddef.h>
@@ -8,8 +9,6 @@
 #include <stdbool.h>
 #include <fcntl.h>
 #include <malloc/malloc.h>
-#include <stdio.h>
-#include <string.h>
 
 typedef struct s_list {
     void *data;
@@ -44,7 +43,7 @@ char *mx_strndup(const char *s1, size_t n);
 char *mx_strncpy(char *dst, const char *src, int len);
 char *mx_strcat(char *restrict s1, const char *restrict s2);
 char *mx_strstr(const char *haystack, const char *needle);
-int mx_strncmp(const char *s1, const char *s2, int n);
+int mx_strncmp(const char *s1, const char *s2, size_t n);
 int mx_get_substr_index(const char *str, const char *sub);
 int mx_count_substr(const char *str, const char *sub);
 int mx_count_words(const char *str, char c);
